+++ Sources/FileGenerators/CMsgSerialiseGenerator.swift
@@ -46,7 +46,7 @@ final public class CMsgSerialiseGenerator: FileGenerator {
 #include \"guwhiteboardserialisation.h\"
 #include \"guwhiteboard_c_types.h\"
 
-size_t serialisemsg(WBTypes message_index, const void *message_in, void *serialised_out)
+int32_t serialisemsg(WBTypes message_index, const void *message_in, void *serialised_out)
 {
     switch (message_index)
     {
